src/main/java/com/lqy/sell/service/impl/ProductInfoServiceImpl.java
package com.lqy.sell.service.impl;

import com.lqy.sell.dataobject.ProductInfo;
import com.lqy.sell.enums.ProductStatusEnum;
import com.lqy.sell.repository.ProductInfoRepository;
import com.lqy.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 商品Service实现
 * Created by Rodriguez
 * 2018/7/12 12:40
 */
@Service
public class ProductInfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoRepository repository;

    @Override
    public ProductInfo findOne(String productInfoId) {
        return repository.findOne(productInfoId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }
}
